Add Max-Forwards header to CANCEL and ACK requests
Canceling an invite request with `clientTransaction.Cancel()` and the ACK sent
automatically to a INVITE response lack a Max-Forwards header. This means some
clients (eg. Linphone) won't accept the request (in fact Linphone rejects the
ACK with a 400, causing a loop of 400 / ACKs between gosip and linphone).

This adds the header to these requests.